ds-nn-m1/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
starlette==0.36.0

# AWS
boto3==1.34.0
botocore==1.34.0

# File type detection
python-magic-bin==0.4.14  # For Windows
# python-magic==0.4.27    # Uncomment for Linux/Mac

# Environment variables
python-dotenv==1.0.0

# API tools
requests==2.31.0

# Utils
typing-extensions==4.9.0
python-multipart==0.0.6